split code into BookingDetailsCard and BookingDetailsWrapper
