+++ src/bookingDetails/ButtonsBar.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { CompleteRideWindow } from 'src/booking-list/components/CompleteRideWindow';
 import { ButtonUI } from 'src/ui-kit/ButtonUI';
-import { DeleteButton } from '../../redux/deleteButton';
+import { DeleteButton } from 'src/redux/deleteButton';
 
 export interface IButtonsBar {
   startTimeSec: string;
