+++ src/bookingDetails/types.ts
@@ -1,0 +1,9 @@
+export interface IDataForBookingDetailPage {
+  firstname: string;
+  carId: string;
+  startTime: string;
+  endTime: string;
+  description: string;
+  isCompleted: string;
+  carLocation: string;
+}
