+++ src/router/root.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter } from 'react-router-dom';
 import LoginForm from 'src/auth/components/loginForm/LoginForm';
 import { App } from 'src/App';
 import { SignUpForm } from 'src/auth/components/signUpForm/SignUpForm';
-import { BookingDetails } from 'src/booking-list/components/BookingDetails';
+import { BookingDetailsWrapper } from 'src/bookingDetails/BookingDetailsWrapper';
 import { BookingList } from 'src/booking-list/components/BookingList';
 import { WrapperForCreatingBooking } from 'src/createNewBooking/components/WrapperForCreatingBooking';
 import {
@@ -30,7 +30,7 @@ const router = createBrowserRouter([
       },
       {
         path: '/booking-details/:bookingId',
-        element: <BookingDetails />,
+        element: <BookingDetailsWrapper />,
       },
       {
         path: creatingNewBooking,
